Add Primitive Operators
This adds operator overloads as extension functions between various IonElement
subtypes and primitive types. The purpose is to improve the usability of
working directly with IonElement values rather than serializing to/from
data classes. Specifically, the following operators are
supported:

* Arithmetic operators and comparisons (+, -, *, /, <, >, >=, <=) between
    * IntElement and IntElement, Long, Int BigInteger
    * DecimalElement and DecimalElement, Double BigDecimal
    * FloatElement and FloatElement, Double
* Boolean infix functions `and` and `or` between BoolElement and Boolean
* `unaryMinus`, `inc` and `dec` for numeric IonElement types

For mixed-type operators, the primitive value can be on the left or right
and the result will always be an IonElement value. For example,
`ionInt(1) + 2 == ionInt(3) == 2 + ionInt(1)`.