Change MutableStructFields to Implement MutableCollection
Implementing MutableCollection instead of Iterator will allow extra
functionality built into MutableCollection. To simplify implementing
`iterator()` on MutableStructFields the internal backing field was also changed
to MutableMap<String, MutableList<StructField>> at the expense of slightly more
overhead because the underlying struct's field name is stored in the map key as
well as value. This can be optimized in the future if desired.